feat(repo): exposed selected from fuel-data-parser
