+++ benchs/data-parser/benches/serialize.rs
@@ -1,10 +1,7 @@
 use async_compression::Level;
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
-use fuel_data_parser::{
-    builder::DataParserBuilder,
-    generate_test_block,
-    types::{CompressionType, SerializationType},
-};
+use data_parser::generate_test_block;
+use fuel_data_parser::{CompressionType, DataParserBuilder, SerializationType};
 use strum::IntoEnumIterator;
 
 fn bench_serialize(c: &mut Criterion) {
